+++ frontend/package.json
@@ -8,9 +8,11 @@
     "@testing-library/user-event": "^13.5.0",
     "autoprefixer": "^10.4.19",
     "axios": "^1.6.8",
+    "date-fns": "^3.6.0",
     "framer-motion": "^11.0.24",
     "postcss": "^8.4.38",
     "react": "^18.2.0",
+    "react-datepicker": "^7.0.0",
     "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
     "react-router-dom": "^6.22.3",
