+++ frontend/src/components/Calendar/DatePicker.css
@@ -1,0 +1,39 @@
+.react-datepicker {
+    border-radius: 10px; 
+    border-radius: 10px !important;
+    border: 2px solid black;
+}
+
+.react-datepicker__header {
+    border-bottom: none;
+    border-top-right-radius: 10px !important; 
+    border-top-left-radius: 10px !important; 
+    background-color: #FEEBA4;
+}
+
+.react-datepicker__day--outside-month {
+    cursor: default;
+    visibility: hidden;
+}
+
+.react-datepicker__month-container {
+    font-size: 1.1rem; 
+    width: 440px;
+}
+
+.react-datepicker__day-name,
+.react-datepicker__day,
+.react-datepicker__time-name {
+    width: 3em; 
+    line-height: 2.5em; 
+}
+
+.react-datepicker__day--selected,
+.react-datepicker__day--keyboard-selected {
+    background-color: #FEEBA4; 
+    color: black;
+}
+
+.react-datepicker__day:hover {
+    background-color: #FEEBA4; 
+}
